ref: move sourcebundle to debug info
