--- sources/Cargo.toml
@@ -1,32 +1,0 @@
-[package]
-name = "symbolic-sources"
-version = "6.1.3"
-license = "MIT"
-authors = [
-    "Armin Ronacher <armin.ronacher@active-4.com>",
-    "Jan Michael Auer <mail@jauer.org>",
-]
-documentation = "https://docs.rs/symbolic-sources"
-homepage = "https://github.com/getsentry/symbolic"
-repository = "https://github.com/getsentry/symbolic"
-description = """
-A library to bundle sources from debug files for later processing.
-"""
-edition = "2018"
-
-exclude = [
-    "tests/**/*",
-]
-
-[dependencies]
-failure = "0.1.5"
-lazy_static = "1.3.0"
-regex = "1.1.6"
-serde = { version = "1.0.91", features = ["derive"] }
-serde_json = "1.0.39"
-symbolic-common = { version = "6.1.3", path = "../common" }
-symbolic-debuginfo = { version = "6.1.3", path = "../debuginfo" }
-zip = "0.5.2"
-
-[dev-dependencies]
-tempfile = "3.0.8"
